feat: restore calc js implementation
